+++ python/graphstorm/gconstruct/construct_graph.py
@@ -36,148 +36,10 @@ import dgl
 from ..utils import sys_tracker
 from .file_io import parse_node_file_format, parse_edge_file_format
 from .file_io import get_in_files
-
-def process_features(data, ops):
-    """ Process the data with the specified operations.
-
-    This function runs the input operations on the corresponding data
-    and returns the processed results.
-
-    Parameters
-    ----------
-    data : dict
-        The data stored as a dict.
-    ops : list of tuples
-        The operations. Each tuple contains two elements. The first element
-        is the data name and the second element is a Python function
-        to process the data.
-
-    Returns
-    -------
-    dict : the key is the data name, the value is the processed data.
-    """
-    new_data = {}
-    for feat_col, feat_name, dtype, op in ops:
-        # If the transformation is defined on the feature.
-        if op is not None:
-            res = op(data[feat_col])
-            if isinstance(res, dict):
-                for key, val in res.items():
-                    new_data[key] = val
-            else:
-                new_data[feat_name] = res
-        # If the required data type is defined on the feature.
-        elif dtype is not None:
-            new_data[feat_name] = data[feat_col].astype(dtype)
-        # If no transformation is defined for the feature.
-        else:
-            new_data[feat_name] = data[feat_col]
-    return new_data
-
-def get_valid_label_index(label):
-    """ Get the index of the samples with valid labels.
-
-    Some of the samples may not have labels. We require users to use
-    NaN to indicate the invalid labels.
-
-    Parameters
-    ----------
-    label : Numpy array
-        The labels of the samples.
-
-    Returns
-    -------
-    Numpy array : the index of the samples with valid labels in the list.
-    """
-    if np.issubdtype(label.dtype, np.floating):
-        return np.logical_not(np.isnan(label)).nonzero()[0]
-    elif np.issubdtype(label.dtype, np.integer):
-        return np.arange(len(label))
-    else:
-        raise ValueError("GraphStorm only supports label data of integers and float." + \
-                         f"This label data has data type of {label.dtype}.")
-
-def process_labels(data, label_confs, is_node):
-    """ Process labels
-
-    Parameters
-    ----------
-    data : dict
-        The data stored as a dict.
-    label_conf : list of dict
-        The list of configs to construct labels.
-    is_node : bool
-        Whether or not to process labels on nodes.
-
-    Returns
-    -------
-    dict of tensors : labels (optional) and train/validation/test masks.
-    """
-    assert len(label_confs) == 1, "We only support one label per node/edge type."
-    label_conf = label_confs[0]
-    assert 'task_type' in label_conf, "'task_type' must be defined in the label field."
-    if label_conf['task_type'] == 'classification':
-        assert 'label_col' in label_conf, \
-                "'label_col' must be defined in the label field."
-        label_col = label_conf['label_col']
-        label = data[label_col]
-        assert np.issubdtype(label.dtype, np.integer) \
-                or np.issubdtype(label.dtype, np.floating), \
-                "The labels for classification have to be integers."
-        valid_label_idx = get_valid_label_index(label)
-        label = np.int32(label)
-        num_samples = len(label)
-    elif label_conf['task_type'] == 'regression':
-        assert 'label_col' in label_conf, \
-                "'label_col' must be defined in the label field."
-        label_col = label_conf['label_col']
-        label = data[label_col]
-        valid_label_idx = get_valid_label_index(label)
-        num_samples = len(label)
-    else:
-        assert label_conf['task_type'] == 'link_prediction', \
-                "The task type must be classification, regression or link_prediction."
-        assert not is_node, "link_prediction task must be defined on edges."
-        label_col = label = None
-        valid_label_idx = None
-        # Any column in the data can define the number of samples in the data.
-        for val in data.values():
-            num_samples = len(val)
-            break
-
-    if 'split_pct' in label_conf:
-        train_split, val_split, test_split = label_conf['split_pct']
-        assert train_split + val_split + test_split <= 1, \
-                "The data split of training/val/test cannot be more than the entire dataset."
-        if valid_label_idx is None:
-            rand_idx = np.random.permutation(num_samples)
-        else:
-            rand_idx = np.random.permutation(valid_label_idx)
-        num_labels = len(rand_idx)
-        num_train = int(num_labels * train_split)
-        num_val = int(num_labels * val_split)
-        num_test = int(num_labels * test_split)
-        val_start = num_train
-        val_end = num_train + num_val
-        test_end = num_train + num_val + num_test
-        train_idx = rand_idx[0:num_train]
-        val_idx = rand_idx[val_start:val_end]
-        test_idx = rand_idx[val_end:test_end]
-        train_mask = np.zeros((num_samples,), dtype=np.int8)
-        val_mask = np.zeros((num_samples,), dtype=np.int8)
-        test_mask = np.zeros((num_samples,), dtype=np.int8)
-        train_mask[train_idx] = 1
-        val_mask[val_idx] = 1
-        test_mask[test_idx] = 1
-    if label_col is None:
-        return {'train_mask': train_mask,
-                'val_mask': val_mask,
-                'test_mask': test_mask}
-    else:
-        return {label_col: label,
-                'train_mask': train_mask,
-                'val_mask': val_mask,
-                'test_mask': test_mask}
+from .transform import parse_feat_ops, process_features
+from .transform import process_labels
+from .id_map import NoopMap, IdMap, map_node_ids
+from .utils import WorkerPool, ExtMemArrayConverter, partition_graph
 
 def parse_node_data(in_file, feat_ops, node_id_col, label_conf, read_file):
     """ Parse node data.
@@ -210,58 +72,6 @@ def parse_node_data(in_file, feat_ops, node_id_col, label_conf, read_file):
         for key, val in label_data.items():
             feat_data[key] = val
     return (data[node_id_col], feat_data)
-
-def map_node_ids(src_ids, dst_ids, edge_type, node_id_map, skip_nonexist_edges):
-    """ Map node IDs of source and destination nodes of edges.
-
-    In the ID mapping, we need to handle multiple errors in the input data:
-    1) we handle the case that endpoint nodes of edges don't exist;
-    2) we handle the case that the data type of node IDs of the endpoint nodes don't
-    match the data type of the keys of the ID map.
-
-    Parameters
-    ----------
-    src_ids : tensor
-        The source nodes.
-    dst_ids : tensor
-        The destination nodes.
-    edge_type : tuple
-        It contains source node type, relation type, destination node type.
-    node_id_map : dict
-        The key is the node type and value is IdMap or NoopMap.
-    skip_nonexist_edges : bool
-        Whether or not to skip edges whose endpoint nodes don't exist.
-
-    Returns
-    -------
-    tuple of tensors : the remapped source and destination node IDs.
-    """
-    src_type, _, dst_type = edge_type
-    new_src_ids, orig_locs = node_id_map[src_type].map_id(src_ids)
-    # If some of the source nodes don't exist in the node set.
-    if len(orig_locs) != len(src_ids):
-        bool_mask = np.ones(len(src_ids), dtype=bool)
-        bool_mask[orig_locs] = False
-        if skip_nonexist_edges:
-            print(f"source nodes of {src_type} do not exist: {src_ids[bool_mask]}")
-        else:
-            raise ValueError(f"source nodes of {src_type} do not exist: {src_ids[bool_mask]}")
-        dst_ids = dst_ids[orig_locs]
-    src_ids = new_src_ids
-
-    new_dst_ids, orig_locs = node_id_map[dst_type].map_id(dst_ids)
-    # If some of the dest nodes don't exist in the node set.
-    if len(orig_locs) != len(dst_ids):
-        bool_mask = np.ones(len(dst_ids), dtype=bool)
-        bool_mask[orig_locs] = False
-        if skip_nonexist_edges:
-            print(f"dest nodes of {dst_type} do not exist: {dst_ids[bool_mask]}")
-        else:
-            raise ValueError(f"dest nodes of {dst_type} do not exist: {dst_ids[bool_mask]}")
-        # We need to remove the source nodes as well.
-        src_ids = src_ids[orig_locs]
-    dst_ids = new_dst_ids
-    return src_ids, dst_ids
 
 def parse_edge_data(in_file, feat_ops, node_id_map, read_file, conf, skip_nonexist_edges):
     """ Parse edge data.
@@ -573,80 +383,6 @@ def verify_confs(confs):
                 f"source node type {src_type} does not exist. Please check your input data."
         assert dst_type in ntypes, \
                 f"dest node type {dst_type} does not exist. Please check your input data."
-
-def partition_graph(g, node_data, edge_data, graph_name, num_partitions, output_dir,
-                    part_method=None):
-    """ Partition a graph
-
-    This takes advantage of the graph partition function in DGL.
-    To save memory consumption for graph partition. We only pass the graph object
-    with the graph structure to DGL's graph partition function.
-    We will split the node/edge feature tensors based on the graph partition results.
-    By doing so, we can keep the node/edge features in external memory to further
-    save memory.
-
-    Parameters
-    ----------
-    g : DGLGraph
-        The full graph object.
-    node_data : dict of tensors
-        The node feature tensors.
-    edge_data : dict of tensors
-        The edge feature tensors.
-    graph_name : str
-        The graph name.
-    num_partitions : int
-        The number of partitions.
-    output_dir : str
-        The directory where we will save the partitioned results.
-    part_method : str (optional)
-        The partition algorithm used to partition the graph.
-    """
-    from dgl.distributed.graph_partition_book import _etype_tuple_to_str
-    orig_id_name = "__gs_orig_id"
-    for ntype in g.ntypes:
-        g.nodes[ntype].data[orig_id_name] = th.arange(g.number_of_nodes(ntype))
-    for etype in g.canonical_etypes:
-        g.edges[etype].data[orig_id_name] = th.arange(g.number_of_edges(etype))
-    sys_tracker.check('Before partitioning starts')
-    if part_method is None:
-        part_method = "None" if num_partitions == 1 else "metis"
-    dgl.distributed.partition_graph(g, graph_name, num_partitions, output_dir,
-                                    part_method=part_method,
-                                    # TODO(zhengda) we need to enable balancing node types.
-                                    balance_ntypes=None,
-                                    balance_edges=True)
-    sys_tracker.check('Graph partitioning')
-    for i in range(num_partitions):
-        part_dir = os.path.join(output_dir, "part" + str(i))
-        data = dgl.data.utils.load_tensors(os.path.join(part_dir, "node_feat.dgl"))
-        # Get the node features for the partition and save the node features in node_feat.dgl.
-        for ntype in node_data:
-            # We store the original node IDs as a node feature when we partition the graph.
-            # We can get the original node IDs from the node features and now
-            # we use them to retrieve the right node features.
-            orig_ids = data[ntype + "/" + orig_id_name]
-            for name, ndata in node_data[ntype].items():
-                data[ntype + "/" + name] = th.tensor(ndata[orig_ids])
-            sys_tracker.check(f'Get node data of node {ntype} in partition {i}')
-        # Delete the original node IDs from the node data.
-        for ntype in g.ntypes:
-            del data[ntype + "/" + orig_id_name]
-        dgl.data.utils.save_tensors(os.path.join(part_dir, "node_feat.dgl"), data)
-
-        data = dgl.data.utils.load_tensors(os.path.join(part_dir, "edge_feat.dgl"))
-        # Get the edge features for the partition and save the edge features in edge_feat.dgl.
-        for etype in edge_data:
-            # We store the original edge IDs as a edge feature when we partition the graph.
-            # We can get the original edge IDs from the edge features and now
-            # we use them to retrieve the right edge features.
-            orig_ids = data[_etype_tuple_to_str(etype) + '/' + orig_id_name]
-            for name, edata in edge_data[etype].items():
-                data[_etype_tuple_to_str(etype) + "/" + name] = th.tensor(edata[orig_ids])
-            sys_tracker.check(f'Get edge data of edge {etype} in partition {i}')
-        for etype in g.canonical_etypes:
-            del data[_etype_tuple_to_str(etype) + '/' + orig_id_name]
-        dgl.data.utils.save_tensors(os.path.join(part_dir, "edge_feat.dgl"), data)
 
 def process_graph(args):
     """ Process the graph.
