+++ python/graphstorm/gconstruct/utils.py
@@ -142,3 +142,77 @@ class ExtMemArrayConverter:
         em_arr = np.memmap(tensor_path, arr.dtype, mode="w+", shape=arr.shape)
         em_arr[:] = arr[:]
         return em_arr
+
+def partition_graph(g, node_data, edge_data, graph_name, num_partitions, output_dir,
+                    part_method=None):
+    """ Partition a graph
+
+    This takes advantage of the graph partition function in DGL.
+    To save memory consumption for graph partition. We only pass the graph object
+    with the graph structure to DGL's graph partition function.
+    We will split the node/edge feature tensors based on the graph partition results.
+    By doing so, we can keep the node/edge features in external memory to further
+    save memory.
+
+    Parameters
+    ----------
+    g : DGLGraph
+        The full graph object.
+    node_data : dict of tensors
+        The node feature tensors.
+    edge_data : dict of tensors
+        The edge feature tensors.
+    graph_name : str
+        The graph name.
+    num_partitions : int
+        The number of partitions.
+    output_dir : str
+        The directory where we will save the partitioned results.
+    part_method : str (optional)
+        The partition algorithm used to partition the graph.
+    """
+    from dgl.distributed.graph_partition_book import _etype_tuple_to_str
+    orig_id_name = "__gs_orig_id"
+    for ntype in g.ntypes:
+        g.nodes[ntype].data[orig_id_name] = th.arange(g.number_of_nodes(ntype))
+    for etype in g.canonical_etypes:
+        g.edges[etype].data[orig_id_name] = th.arange(g.number_of_edges(etype))
+    sys_tracker.check('Before partitioning starts')
+    if part_method is None:
+        part_method = "None" if num_partitions == 1 else "metis"
+    dgl.distributed.partition_graph(g, graph_name, num_partitions, output_dir,
+                                    part_method=part_method,
+                                    # TODO(zhengda) we need to enable balancing node types.
+                                    balance_ntypes=None,
+                                    balance_edges=True)
+    sys_tracker.check('Graph partitioning')
+    for i in range(num_partitions):
+        part_dir = os.path.join(output_dir, "part" + str(i))
+        data = dgl.data.utils.load_tensors(os.path.join(part_dir, "node_feat.dgl"))
+        # Get the node features for the partition and save the node features in node_feat.dgl.
+        for ntype in node_data:
+            # We store the original node IDs as a node feature when we partition the graph.
+            # We can get the original node IDs from the node features and now
+            # we use them to retrieve the right node features.
+            orig_ids = data[ntype + "/" + orig_id_name]
+            for name, ndata in node_data[ntype].items():
+                data[ntype + "/" + name] = th.tensor(ndata[orig_ids])
+            sys_tracker.check(f'Get node data of node {ntype} in partition {i}')
+        # Delete the original node IDs from the node data.
+        for ntype in g.ntypes:
+            del data[ntype + "/" + orig_id_name]
+        dgl.data.utils.save_tensors(os.path.join(part_dir, "node_feat.dgl"), data)
+
+        data = dgl.data.utils.load_tensors(os.path.join(part_dir, "edge_feat.dgl"))
+        # Get the edge features for the partition and save the edge features in edge_feat.dgl.
+        for etype in edge_data:
+            # We store the original edge IDs as a edge feature when we partition the graph.
+            # We can get the original edge IDs from the edge features and now
+            # we use them to retrieve the right edge features.
+            orig_ids = data[_etype_tuple_to_str(etype) + '/' + orig_id_name]
+            for name, edata in edge_data[etype].items():
+                data[_etype_tuple_to_str(etype) + "/" + name] = th.tensor(edata[orig_ids])
+            sys_tracker.check(f'Get edge data of edge {etype} in partition {i}')
+        for etype in g.canonical_etypes:
+            del data[_etype_tuple_to_str(etype) + '/' + orig_id_name]
+        dgl.data.utils.save_tensors(os.path.join(part_dir, "edge_feat.dgl"), data)
